Issue #875 Added tri-state for Point fields movable and hidden.

Former-commit-id: 70217e15563e376b86223c42040611a2d1f34582 [formerly 15833c5ca3b8415b3a2fef1da98c42e95721331d]
Former-commit-id: 66d98d068c30564ac76789ac9ec402e1ba52c9e3

+++ edexOsgi/com.raytheon.uf.common.localization/src/com/raytheon/uf/common/localization/LocalizationNotificationObserver.java
@@ -47,6 +47,8 @@ import com.raytheon.uf.common.status.UFStatus.Priority;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Apr 8, 2011            mschenke     Initial creation
+ * Sep 17, 2012 #875       rferrel     Added check that file exists before 
+ *                                      deleting it.
  * 
  * </pre>
  * 
@@ -196,7 +198,7 @@ public class LocalizationNotificationObserver {
         // If file deleted, delete from filesystem if non directory
         if (fum.getChangeType() == FileChangeType.DELETED) {
             File local = pm.adapter.getPath(fum.getContext(), filename);
-            if (local != null && local.isDirectory() == false) {
+            if (local != null && local.isDirectory() == false && local.exists()) {
                 local.delete();
             }
         }
